pyproject.toml
[project]
name = "emotegan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "equinox>=0.11.5",
    "jax>=0.4.31",
    "opencv-python>=4.10.0.84",
    "optax>=0.2.3",
    "pillow>=10.4.0",
    "requests>=2.32.3",
    "selenium>=4.24.0",
    "torchvision>=0.19.0",
]
